// src/schemas/authSchema.ts
import { z } from 'zod';

export const loginSchema = z.object({
  email: z.string().email('Invalid email address'),
  password: z.string().min(6, 'Password must be at least 6 characters long'),
});

export const registerSchema = z.object(
  {
    email: z.string().email('Invalid email address'),
    name: z.string().min(2, 'Name must be at least 2 characters long'),
    password: z.string().min(8, 'Password must be at least 6 characters long'),
  }
)